Extract DrawerViewModel.toState() to consolidate state construction logic
